[batching] minor refactor for BlazingHostTable
